Extracted image generation logic to a separate file.
